pyproject.toml
# pyproject.toml - Modern Python Project Configuration

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "factory-part-recognition"
version = "1.0.0"
description = "CNN-based classification of industrial parts with FastAPI backend"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
keywords = ["machine-learning", "computer-vision", "pytorch", "image-classification", "industrial"]
requires-python = ">=3.10"

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "sqlalchemy>=2.0.0",
    "pillow>=9.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.2.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "ipywidgets>=8.0.0",
]
mlops = [
    "mlflow>=2.0.0",
    "dvc>=3.0.0",
    "wandb>=0.15.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/factory-part-recognition"
Documentation = "https://github.com/yourusername/factory-part-recognition#readme"
Repository = "https://github.com/yourusername/factory-part-recognition.git"
Issues = "https://github.com/yourusername/factory-part-recognition/issues"

[project.scripts]
factory-train = "src.train_improved:train"
factory-evaluate = "src.evaluate:run_evaluation"
factory-serve = "main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

# Ruff Configuration (Linting + Formatting)
[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

# Black Configuration
[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

# MyPy Configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# Pytest Configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=src --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = ["."]

# Coverage Configuration
[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
